refactor: 폴더 병합과 컴포넌트 이름 수정
