--- frontend/src/components/Members/MembersTitle/MyPageTitle.stories.tsx
@@ -1,18 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import MembersTitle from './MembersTitle';
-
-import member from '@/mocks/data/members.json';
-
-const meta: Meta<typeof MembersTitle> = {
-  title: 'Members/MembersTitle',
-  component: MembersTitle,
-  args: {
-    member: member,
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof MembersTitle>;
-
-export const Default: Story = {};
